Rename LabelWithInfoItems -> InformationItems
(cherry picked from commit bbba20c)
(sort of ... parallel with different directory structure)
