+++ src/components/app-controllers/SingleAppController/SingleAppController.js
@@ -21,7 +21,7 @@ import Selector from '../../Selector';
 import VariableDescriptionSelector from '../../VariableDescriptionSelector';
 import {
   modelSelectorLabel, emissionScenarioSelectorLabel, variableSelectorLabel
-} from '../../help-etc/LabelWithInfoItems';
+} from '../../help-etc/InformationItems';
 
 import AppMixin from '../../AppMixin';
 import g from '../../../core/geo';
